refactor: remove v1 label models, add FlawLabel alias
Delete the old label_v2.py (renamed to label.py via git mv).
Remove v1 models (FlawCollaborator, FlawCollaboratorManager)
that were dead code after the v2 migration.

The Django model class stays FlawLabelV2 (renaming polymorphic
parent models is not supported by Django migrations when child
models have explicit parent_link fields). A public alias
FlawLabel = FlawLabelV2 is provided so all imports use the
clean name. related_name changed from "labels_v2" to "labels".
