+++ osidb/models/flaw/flaw.py
@@ -860,7 +860,7 @@ class Flaw(
         """
         check if the flaw has a label with the given name
         """
-        return self.labels_v2.filter(name=label).exists()
+        return self.labels.filter(name=label).exists()
 
     @property
     def is_placeholder(self):
