+++ osidb/models/flaw/label.py
@@ -1,11 +1,21 @@
-import uuid
-from typing import List
+"""
+Polymorphic label models.
 
+Each label type inherits from FlawLabel and gets its own table
+with type-specific fields. The polymorphic_ctype column auto-tracks
+which subclass each row belongs to.
+"""
+
+import uuid
+
+import pghistory
 from django.contrib.postgres import fields
 from django.contrib.postgres.indexes import GinIndex
 from django.core.exceptions import ValidationError
-from django.db import IntegrityError, models
+from django.db import models
 from django.db.models import Q
+from polymorphic.managers import PolymorphicManager
+from polymorphic.models import PolymorphicModel
 
 from osidb.mixins import (
     ACLMixin,
@@ -15,55 +25,341 @@ from osidb.mixins import (
     ValidateMixin,
     validator,
 )
-from osidb.models import Affect, Flaw
-from osidb.query_sets import CustomQuerySetUpdatedDt
 
 
-class FlawLabelManager(models.Manager):
-    @staticmethod
-    def get_filtered(ps_modules: List[str], ps_components: List[str], *args):
-        """Get product family labels based on the ps_modules and ps_components"""
-        return FlawLabel.objects.filter(
-            type=FlawLabel.FlawLabelType.PRODUCT_FAMILY
-        ).filter(
-            (
-                Q(ps_modules__overlap=list(set(ps_modules)))
-                | Q(ps_components__overlap=list(set(ps_components)))
-            )
-            & ~(
-                Q(ps_modules_exclude__overlap=list(set(ps_modules)))
-                | Q(ps_components_exclude__overlap=list(set(ps_components)))
-            ),
-            *args,
-        )
+class FlawLabelManager(ACLMixinManager, PolymorphicManager, TrackingMixinManager):
+    """Manager combining ACL annotations with polymorphic queries."""
 
 
-class FlawLabel(models.Model):
-    class FlawLabelType(models.TextChoices):
+@pghistory.track(
+    pghistory.InsertEvent(),
+    pghistory.UpdateEvent(),
+    pghistory.DeleteEvent(),
+    model_name="FlawLabelV2Audit",
+)
+class FlawLabelV2(PolymorphicModel, ACLMixin, TrackingMixin, ValidateMixin):
+    """
+    Base polymorphic label model.
+
+    All label types inherit from this model. The polymorphic_ctype field
+    automatically tracks which subclass each instance belongs to.
+    """
+
+    class LabelType(models.TextChoices):
         ALIAS = "alias"
         BU = "bu"
         CONTEXT_BASED = "context_based"
         PRODUCT_FAMILY = "product_family"
         WORKFLOW = "workflow"
 
-    # internal primary key
     uuid = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
 
-    # Label name
+    flaw = models.ForeignKey(
+        "Flaw",
+        on_delete=models.CASCADE,
+        related_name="labels",
+    )
+
+    name = models.CharField(max_length=255)
+
+    objects = FlawLabelManager()
+
+    class Meta:
+        constraints = [
+            models.UniqueConstraint(
+                fields=["flaw", "name"], name="unique_label_per_flaw_v2"
+            )
+        ]
+        indexes = TrackingMixin.Meta.indexes + [
+            models.Index(fields=["flaw"]),
+            models.Index(fields=["name"]),
+            GinIndex(fields=["acl_read"]),
+        ]
+
+    # Subclasses must define their type
+    type = None
+
+    def __str__(self):
+        return f"{self.name} ({self.type})"
+
+    def save(self, *args, **kwargs):
+        self.inherit_parent_flaw_acls()
+        super().save(*args, **kwargs)
+
+    def validate(self):
+        """
+        Run custom validators and standard Django validations.
+
+        This overrides ValidateMixin.validate()
+        to call @validator methods before full_clean().
+        """
+        for validator_name in self._validators:
+            getattr(self, validator_name)()
+        super().validate()
+
+
+class AliasLabel(FlawLabelV2):
+    """
+    Free-form alias labels.
+
+    These labels don't require pre-registration and can be any text.
+    Used for incident IDs, bug tracker references, or any free-form tagging.
+
+    No additional fields needed.
+    """
+
+    type = FlawLabelV2.LabelType.ALIAS
+
+    class Meta:
+        verbose_name = "Alias Label"
+        verbose_name_plural = "Alias Labels"
+
+
+class BULabel(FlawLabelV2):
+    """
+    Business unit labels with state tracking.
+
+    Similar to CollaboratorLabel but for business unit-specific workflows.
+    Requires pre-registration in BULabelDefinition.
+    """
+
+    type = FlawLabelV2.LabelType.BU
+
+    class State(models.TextChoices):
+        NEW = "NEW"
+        REQ = "REQ"
+        SKIP = "SKIP"
+        DONE = "DONE"
+
+    state = models.CharField(
+        max_length=10,
+        choices=State.choices,
+        default=State.NEW,
+    )
+    contributor = models.CharField(max_length=255, blank=True)
+    relevant = models.BooleanField(default=True)
+
+    class Meta:
+        verbose_name = "BU Label"
+        verbose_name_plural = "BU Labels"
+
+    @validator
+    def _validate_pre_registration(self, **kwargs):
+        """Validate that the label name is pre-registered"""
+        if not BULabelDefinition.objects.filter(name=self.name).exists():
+            raise ValidationError(
+                {"name": f"BU label '{self.name}' must be pre-registered."}
+            )
+
+
+class CollaboratorLabel(FlawLabelV2):
+    """
+    Context-based labels with full state management.
+
+    These are manually created labels that require pre-registration
+    in CollaboratorLabelDefinition and support full workflow state tracking.
+    """
+
+    type = FlawLabelV2.LabelType.CONTEXT_BASED
+
+    class State(models.TextChoices):
+        NEW = "NEW"
+        REQ = "REQ"
+        SKIP = "SKIP"
+        DONE = "DONE"
+
+    state = models.CharField(
+        max_length=10,
+        choices=State.choices,
+        default=State.NEW,
+    )
+    contributor = models.CharField(max_length=255, blank=True)
+    relevant = models.BooleanField(default=True)
+
+    class Meta:
+        verbose_name = "Collaborator Label"
+        verbose_name_plural = "Collaborator Labels"
+
+    @validator
+    def _validate_pre_registration(self, **kwargs):
+        """Validate that the label name is pre-registered"""
+        if not CollaboratorLabelDefinition.objects.filter(name=self.name).exists():
+            raise ValidationError(
+                {"name": f"Collaborator label '{self.name}' must be pre-registered."}
+            )
+
+
+class ProductFamilyLabel(FlawLabelV2):
+    """
+    Auto-created labels based on product definitions.
+
+    These labels are automatically created from affects based on
+    ps_module/ps_component matching rules defined in ProductFamilyLabelDefinition.
+    They cannot be manually deleted via API.
+    """
+
+    type = FlawLabelV2.LabelType.PRODUCT_FAMILY
+
+    relevant = models.BooleanField(default=True)
+
+    class Meta:
+        verbose_name = "Product Family Label"
+        verbose_name_plural = "Product Family Labels"
+
+    @staticmethod
+    def create_from_affect(affect):
+        """
+        Add new product family labels to the flaw based on the affect
+        """
+        existing = set(affect.flaw.labels.values_list("name", flat=True))
+
+        definitions = ProductFamilyLabelDefinition.get_matching(
+            [affect.ps_module], [affect.ps_component]
+        )
+
+        for definition in definitions:
+            if definition.name not in existing:
+                ProductFamilyLabel.objects.create(
+                    flaw=affect.flaw,
+                    name=definition.name,
+                    acl_read=list(affect.flaw.acl_read),
+                    acl_write=list(affect.flaw.acl_write),
+                )
+            else:
+                ProductFamilyLabel.objects.filter(
+                    flaw=affect.flaw, name=definition.name, relevant=False
+                ).update(relevant=True)
+
+    @staticmethod
+    def update_relevance(flaw):
+        """
+        Update product family label relevance based on current affects
+        """
+        from osidb.models import Affect
+
+        ps_values = Affect.objects.filter(flaw=flaw).values_list(
+            "ps_module", "ps_component"
+        )
+
+        if not ps_values:
+            current_names = set()
+        else:
+            ps_modules, ps_components = zip(*ps_values)
+            definitions = ProductFamilyLabelDefinition.get_matching(
+                ps_modules, ps_components
+            )
+            current_names = {d.name for d in definitions}
+
+        for label in ProductFamilyLabel.objects.filter(flaw=flaw):
+            new_relevant = label.name in current_names
+            if label.relevant != new_relevant:
+                label.relevant = new_relevant
+                label.save()
+
+            if label.name in current_names:
+                current_names.discard(label.name)
+
+        existing_names = set(flaw.labels.values_list("name", flat=True))
+
+        for name in current_names:
+            if name not in existing_names:
+                ProductFamilyLabel.objects.create(
+                    flaw=flaw,
+                    name=name,
+                    acl_read=list(flaw.acl_read),
+                    acl_write=list(flaw.acl_write),
+                )
+
+
+class WorkflowLabel(FlawLabelV2):
+    """
+    Workflow classification markers.
+
+    These labels don't require pre-registration and represent binary flags.
+    Presence of the label(s) navigates the workflow classification.
+
+    No additional fields needed.
+    """
+
+    type = FlawLabelV2.LabelType.WORKFLOW
+
+    class Meta:
+        verbose_name = "Workflow Label"
+        verbose_name_plural = "Workflow Labels"
+
+
+# Public alias — the rest of the codebase imports FlawLabel.
+# The Django model class stays FlawLabelV2 until the polymorphic
+# RenameModel migration issue is resolved separately.
+FlawLabel = FlawLabelV2
+
+FlawLabelV2.TYPE_TO_MODEL = {
+    FlawLabelV2.LabelType.ALIAS: AliasLabel,
+    FlawLabelV2.LabelType.BU: BULabel,
+    FlawLabelV2.LabelType.CONTEXT_BASED: CollaboratorLabel,
+    FlawLabelV2.LabelType.PRODUCT_FAMILY: ProductFamilyLabel,
+    FlawLabelV2.LabelType.WORKFLOW: WorkflowLabel,
+}
+
+
+# Label Definition Registry
+# Only label types that require pre-registration have definition models.
+# ALIAS and WORKFLOW labels are free-form and have no definition models.
+
+
+class BaseLabelDefinition(models.Model):
+    """
+    Abstract base class for label definitions.
+
+    Provides common fields for all definition types.
+    Only types that require pre-registration inherit from this.
+    """
+
+    uuid = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
     name = models.CharField(max_length=255, unique=True)
 
-    # Label type
-    type = models.CharField(
-        max_length=20,
-        choices=FlawLabelType.choices,
-        editable=False,
-    )
+    class Meta:
+        abstract = True
+
+    def __str__(self):
+        return self.name
+
+
+class BULabelDefinition(BaseLabelDefinition):
+    """
+    Registry for Business Unit labels.
+
+    These labels must be pre-registered before they can be used.
+    No additional fields needed - just name validation.
+    """
+
+    class Meta:
+        verbose_name = "BU Label Definition"
+        verbose_name_plural = "BU Label Definitions"
+
+
+class CollaboratorLabelDefinition(BaseLabelDefinition):
+    """
+    Registry for context-based labels.
+
+    These labels must be pre-registered before they can be used.
+    No additional fields needed - just name validation.
+    """
+
+    class Meta:
+        verbose_name = "Collaborator Label Definition"
+        verbose_name_plural = "Collaborator Label Definitions"
+
+
+class ProductFamilyLabelDefinition(BaseLabelDefinition):
+    """
+    Registry for product family labels with auto-creation filters.
+
+    These labels are automatically created on flaws when affects match
+    the ps_module/ps_component filter criteria defined here.
+    """
 
     ps_components = fields.ArrayField(
-        models.CharField(max_length=100), default=list, blank=True
-    )
-
-    ps_components_exclude = fields.ArrayField(
         models.CharField(max_length=100), default=list, blank=True
     )
 
@@ -71,166 +367,30 @@ class FlawLabel(models.Model):
         models.CharField(max_length=100), default=list, blank=True
     )
 
+    ps_components_exclude = fields.ArrayField(
+        models.CharField(max_length=100), default=list, blank=True
+    )
+
     ps_modules_exclude = fields.ArrayField(
         models.CharField(max_length=100), default=list, blank=True
     )
 
-    objects = FlawLabelManager()
-
-    def __str__(self):
-        return self.name
-
-
-class FlawCollaboratorManager(ACLMixinManager, TrackingMixinManager):
-    """collaborator manager"""
-
-    @staticmethod
-    def create_from_affect(affect: Affect):
-        """Add new labels to the flaw based on the affect"""
-
-        labels = FlawLabel.objects.get_filtered(
-            [affect.ps_module],
-            [affect.ps_component],
-            ~Q(name__in=affect.flaw.labels.values_list("label", flat=True)),
-        )
-
-        for label in labels:
-            FlawCollaborator.objects.create(
-                flaw=affect.flaw,
-                label=label.name,
-                state=FlawCollaborator.FlawCollaboratorState.NEW,
-                type=label.type,
-                acl_read=list(affect.flaw.acl_read),
-                acl_write=list(affect.flaw.acl_write),
-            )
-
-    @staticmethod
-    def create_from_flaw(flaw: Flaw):
-        """Add new labels to the flaw based on the flaw"""
-
-        ps_values = Affect.objects.filter(flaw=flaw).values_list(
-            "ps_module", "ps_component"
-        )
-
-        if not ps_values:
-            return []
-
-        [ps_modules, ps_components] = list(zip(*ps_values))
-        labels = FlawLabel.objects.get_filtered(ps_modules, ps_components)
-
-        for label in labels:
-            FlawCollaborator.objects.get_or_create(
-                flaw=flaw,
-                label=label.name,
-                type=label.type,
-                defaults={
-                    "state": FlawCollaborator.FlawCollaboratorState.NEW,
-                    "acl_read": list(flaw.acl_read),
-                    "acl_write": list(flaw.acl_write),
-                },
-            )
-
-        return labels
-
-    @staticmethod
-    def mark_irrelevant(flaw: Flaw):
-        """Mark labels as irrelevant based on the current affects"""
-
-        collaborators = FlawCollaborator.objects.filter(
-            flaw=flaw, type=FlawLabel.FlawLabelType.PRODUCT_FAMILY
-        )
-        new_labels = [
-            label.name for label in FlawCollaborator.objects.create_from_flaw(flaw)
-        ]
-        for collaborator in collaborators:
-            if collaborator.label not in new_labels:
-                collaborator.relevant = False
-                collaborator.save()
-
-
-class FlawCollaborator(ACLMixin, TrackingMixin, ValidateMixin):
-    class FlawCollaboratorState(models.TextChoices):
-        NEW = "NEW"
-        REQ = "REQ"
-        SKIP = "SKIP"
-        DONE = "DONE"
-
-    # internal primary key
-    uuid = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
-
-    flaw = models.ForeignKey(
-        Flaw,
-        on_delete=models.CASCADE,
-        related_name="labels",
-    )
-
-    label = models.CharField(max_length=255)
-
-    type = models.CharField(
-        max_length=20,
-        choices=FlawLabel.FlawLabelType.choices,
-        editable=False,
-        default=FlawLabel.FlawLabelType.CONTEXT_BASED,
-    )
-
-    state = models.CharField(
-        max_length=10,
-        choices=FlawCollaboratorState.choices,
-        default=FlawCollaboratorState.NEW,
-    )
-
-    contributor = models.CharField(max_length=255, blank=True)
-
-    relevant = models.BooleanField(default=True)
-
-    objects = FlawCollaboratorManager.from_queryset(CustomQuerySetUpdatedDt)()
-
     class Meta:
-        constraints = [
-            models.UniqueConstraint(
-                fields=["flaw", "label"], name="unique label per flaw"
-            ),
-        ]
-        indexes = TrackingMixin.Meta.indexes + [
-            GinIndex(fields=["acl_read"]),
-        ]
+        verbose_name = "Product Family Label Definition"
+        verbose_name_plural = "Product Family Label Definitions"
 
-    def save(self, *args, **kwargs):
-        # Inherit parent Flaw ACLs when unset so ORM creates stay RLS-safe.
-        self.inherit_parent_flaw_acls()
-
-        # Workflow labels are always "done" and "relevant" - they're binary flags
-        if self.type == FlawLabel.FlawLabelType.WORKFLOW:
-            self.state = FlawCollaborator.FlawCollaboratorState.DONE
-            self.relevant = True
-
-        try:
-            super().save(*args, **kwargs)
-        except IntegrityError as e:
-            ex_msg = str(e)
-            if "duplicate key value violates unique constraint" in ex_msg:
-                raise ValidationError(
-                    {"label": f"Label '{self.label}' already exists."}
-                )
-
-    def validate(self):
+    @staticmethod
+    def get_matching(ps_modules, ps_components):
         """
-        Run ACLMixin @validator methods before standard Django validations.
-        (AlertMixin is heavier than label needs; same pattern as FlawLabelV2.)
+        Get definitions matching given ps_modules and ps_components
         """
-        for validator_name in self._validators:
-            getattr(self, validator_name)()
-        super().validate()
-
-    @validator
-    def _validate_label(self, **kwargs):
-        """Validate the label"""
-        # Alias and workflow labels don't need pre-registration
-        if self.type in (
-            FlawLabel.FlawLabelType.ALIAS,
-            FlawLabel.FlawLabelType.WORKFLOW,
-        ):
-            return
-
-        if not FlawLabel.objects.filter(name=self.label).exists():
-            raise ValidationError({"label": f"Label '{self.label}' does not exist."})
+        return ProductFamilyLabelDefinition.objects.filter(
+            (
+                Q(ps_modules__overlap=list(set(ps_modules)))
+                | Q(ps_components__overlap=list(set(ps_components)))
+            )
+            & ~(
+                Q(ps_modules_exclude__overlap=list(set(ps_modules)))
+                | Q(ps_components_exclude__overlap=list(set(ps_components)))
+            )
+        )
